refactor almost all the commands into plugins
