+++ package.json
@@ -11,25 +11,15 @@
   },
   "license": "GPL-2.0",
   "dependencies": {
-    "d20": "^1.4.0",
     "discord.js": "11.0.x",
     "feedparser": "1.1.x",
-    "google": "1.1.x",
-    "google-images": "0.1.x",
     "html-to-text": "1.5.x",
-    "imgflipper": "1.0.x",
-    "leet": "^1.2.7",
     "nock": "1.7.x",
-    "node-wolfram": "0.0.1",
     "npm": "",
     "querystring": "0.2.x",
     "sinon": "1.14.x",
     "tumblr.js": "0.0.x",
-    "underscore": "1.8.x",
-    "urban": "^0.3.0",
-    "wikijs": "0.1.x",
-    "yahoo-finance": "0.2.x",
-    "youtube-node": "1.2.x"
+    "underscore": "1.8.x"
   },
   "main": "discord_bot.js"
 }
