+++ plugins/Admin/package.json
@@ -1,0 +1,6 @@
+{
+  "main": "admin.js",
+  "dependencies": {
+    
+  }
+}
