+++ plugins/Channel Management/package.json
@@ -1,0 +1,6 @@
+{
+  "main": "channels.js",
+  "dependencies": {
+    
+  }
+}
