+++ plugins/Dev/package.json
@@ -1,0 +1,6 @@
+{
+  "main": "dev.js",
+  "dependencies": {
+    
+  }
+}
