+++ plugins/Google/package.json
@@ -1,0 +1,7 @@
+{
+  "main": "google.js",
+  "dependencies": {
+    "request": ">=2.58.0",
+	"youtube-node": "1.2.x"
+  }
+}
