--- plugins/Images/package.json
@@ -1,6 +1,0 @@
-{
-  "main": "images.js",
-  "dependencies": {
-    "request": ">=2.58.0"
-  }
-}
