+++ plugins/Imgflip/package.json
@@ -1,0 +1,6 @@
+{
+  "main": "imgflip.js",
+  "dependencies": {
+    "imgflipper": "1.0.x"
+  }
+}
