+++ plugins/Misc/package.json
@@ -1,0 +1,6 @@
+{
+  "main": "misc.js",
+  "dependencies": {
+    "request": "2.x.x"
+  }
+}
