+++ plugins/RSS/package.json
@@ -1,0 +1,6 @@
+{
+  "main": "rss.js",
+  "dependencies": {
+    
+  }
+}
