+++ plugins/Urban Dictionary/package.json
@@ -1,0 +1,6 @@
+{
+  "main": "urban.js",
+  "dependencies": {
+    "urban": "^0.3.0"
+  }
+}
