+++ plugins/Wikipedia/package.json
@@ -1,0 +1,6 @@
+{
+  "main": "wikipedia.js",
+  "dependencies": {
+    "wikijs": "0.1.x"
+  }
+}
