+++ plugins/leet/leet.js
@@ -1,0 +1,13 @@
+exports.commands = [
+	"leet"
+]
+
+var leet = require("leet");
+
+exports.leet = {
+	usage: "<message>",
+	description: "converts boring regular text to 1337",
+	process: function(bot,msg,suffix){
+		msg.channel.sendMessage(leet.convert(suffix));
+	}
+}
