+++ plugins/xkcd/xkcd.json
@@ -1,0 +1,6 @@
+{
+  "main": "cleverbot.js",
+  "dependencies": {
+    "cleverbot-node": "0.2.x"
+  }
+}
